requirements.txt
# Multi-Agent Orchestration System - Clean Dependencies
# Generated: 2025-11-25

chromadb==1.3.5
sentence-transformers==5.1.2

# LangChain Core
langchain==0.0.350
langchain-community==0.0.20
langchain-core==1.1.0
langchain-text-splitters==0.3.9

# LangGraph
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langsmith==0.4.46

# Google AI
google-generativeai==0.8.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.34.0
langchain-google-genai==3.1.0

# Optional: Vertex AI
langchain-google-vertexai==3.0.3
google-cloud-aiplatform==1.128.0

# Web Framework
streamlit==1.49.1
fastapi==0.104.1
starlette==0.27.0
uvicorn==0.24.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3

# Observability
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
